requirements.txt
asgiref==3.2.10
Django==3.1.1
django-environ==0.4.5
django-versatileimagefield==2.0
djangorestframework==3.11.1
drf-extensions==0.6.0
Pillow==7.2.0
python-magic==0.4.18
pytz==2020.1
sqlparse==0.3.1
